src/views: new EditRule view
